Update README.md
There are many different ways to approach the Titanic survival prediction problem. One common approach is to use a logistic regression model. Logistic regression is a statistical model that can be used to predict the probability of an event occurring. In this case, the event is whether or not a passenger survived the Titanic.

To train a logistic regression model, we need to first collect a data set of passengers and their survival status. We can then use this data set to fit the model. Once the model is fit, we can use it to predict the survival status of new passengers.

Another approach to the Titanic survival prediction problem is to use a decision tree model. Decision trees are a type of machine learning model that can be used to make decisions. In this case, the decision is whether or not a passenger survived the Titanic.

To train a decision tree model, we need to first collect a data set of passengers and their survival status. We can then use this data set to fit the model. Once the model is fit, we can use it to predict the survival status of new passengers.

There are many other approaches to the Titanic survival prediction problem. The best approach for a particular problem will depend on the data set and the desired outcome.